feat(单元测试): 添加Service层单元测试
- 新增84个单元测试覆盖7个Service类
  - TourismScenicSpotService: 13个测试
  - TourismActivityService: 15个测试
  - TourismVenueService: 18个测试
  - TourismImageService: 13个测试
  - TourismActivityApprovalService: 14个测试
  - TourismNewsService: 6个测试
  - TourismNoticeService: 5个测试
- 修复ApprovalServiceImpl中LambdaUpdateWrapper导致测试失败问题
- 优化pom.xml依赖配置

+++ cquet-tourism/src/main/java/cn/edu/cquet/tourism/domain/vo/TourismScenicSpotQueryVo.java
@@ -1,5 +1,7 @@
 package cn.edu.cquet.tourism.domain.vo;
 
+import java.math.BigDecimal;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
@@ -20,10 +22,10 @@ public class TourismScenicSpotQueryVo {
     private String level;
 
     @Schema(description = "最低门票价格")
-    private Double minTicketPrice;
+    private BigDecimal minTicketPrice;
 
     @Schema(description = "最高门票价格")
-    private Double maxTicketPrice;
+    private BigDecimal maxTicketPrice;
 
     @Schema(description = "景区状态(0:关闭, 1:开放)")
     private String status;
